Combo Release (v4)
Stupid things I wrote. Including this website, now with major changes and lifetime improvements. Combining updates that haven’t been released (from 3.3 and 3.4)  

- Added Landing Page
- Added About Page (WIP, haven't linked yet)
- You can now access the website without specifying the .html file - Reorganized Everything, from File management to reducing unecessary code.
- Under the hood improvements

Not just that, we also fixed: 
- An Issue where website content doesn’t fit to user’s screen.
- An Issue where background images also moves when scrolling on Safari Browser.
- An Issue where website zoom is an issue to larger phones.

Thanks for standing by! We will add more updates and features to the site. Current features/projects we pitched in:  

- Secure Password Generator Homepage
- What the Void?! - Price Calculator/iOS calculator clone
- Knowledge Center
- Homer’s Webpage  

They are currently in the lowest priority now, but rest assured we will add them in a couple of days. Mobile Data is a scarce resource right now.

Thanks for understanding!